Add type checking
This adds it through Tox and in the GitHub Checks CI. This adds both pyright
and mypy, but only pyright is used in the CI on GitHub, since it's faster.

We have *a lot* of errors/warnings at the moment. It's good to get this in now so
that people can start using it to find and fix those errors/warnings.